requirements.txt
discord.py
python-dotenv
flask
flask-restful
starcoin-sdk-python
flask-cors